Website Backend & Frontend/smartHome/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asgiref==3.11.1
assemblyai==0.50.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.4
daphne==4.2.1
distro==1.9.0
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-genai==1.62.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jiter==0.13.0
msgpack==1.1.2
openai==2.17.0
packaging==26.0
pillow==12.1.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dotenv==1.2.1
python-telegram-bot==22.6
redis==7.1.0
requests==2.32.5
service-identity==24.2.0
singbox2proxy==0.2.7
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.3
sqlparse==0.5.5
tenacity==9.1.3
tomli==2.4.0
tqdm==4.67.3
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
zope.interface==8.2
